idss-mcp/mcp-server/requirements.txt
# MCP E-commerce Server - Stage 1 Dependencies
# Python 3.11+ recommended

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache
redis==5.0.1
hiredis==2.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Vector Search & Semantic Matching (Optional but recommended)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Knowledge Graph (Stage 3A - per week4notes.txt)
neo4j>=5.15.0

# Web Scraping (for real product data)
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0  # For dynamic sites like Shopify
lxml>=4.9.0  # Faster HTML parsing
pandas>=2.0.0  # For ShopifyScraper package

# LLM (interview question generator: laptop/vehicle follow-up questions)
openai>=1.0.0

# Utilities
python-dotenv==1.0.0
